back-end/AI/requirements.txt


fastapi==0.115.0
uvicorn==0.30.6
pymongo==4.8.0
sentence-transformers==3.1.1
python-dotenv==1.0.1
numpy==1.26.4